+++ src/modules/products/product.route.ts
