+++ src/server.ts
@@ -2,12 +2,12 @@ import Fastify, { FastifyRequest, FastifyReply } from "fastify";
 import fjwt, { JWT } from "@fastify/jwt";
 import fs from "fs";
 import pino from "pino";
-import { userSchemas } from "./modules/user/user.schema";
-import userRoutes from "./modules/user/user.route";
-import productRoutes from "./modules/product/product.route";
-import { productSchemas } from "./modules/product/product.schema";
-import orderRoutes from "./modules/order/order.route";
-import { orderSchemas } from "./modules/order/order.schema";
+import { userSchemas } from "./modules/users/user.schema";
+import userRoutes from "./modules/users/user.route";
+import productRoutes from "./modules/products/product.route";
+import { productSchemas } from "./modules/products/product.schema";
+import orderRoutes from "./modules/orders/order.route";
+import { orderSchemas } from "./modules/orders/order.schema";
 import stockRoutes from "./modules/stocks/stock.route";
 import { stockSchemas } from "./modules/stocks/stock.schema";
 import { UserRole } from "@prisma/client";
